Rename `ObjectInspector.inspect` to `ObjectFormatter.format`.
`inspect` is a ruby protocol, which is expected to be callable
with no arguments.  IRB, for example, will call `inspect` on
any object returned by an expression. Having `ObjectInspector.inspect` require an argument is problematic
because it violated that standard.
